Update project structure and dependencies; move styles to modules, upgrade Next.js version, and clean up TODOs
