+++ TODO.txt
@@ -1,8 +1,5 @@
-Fix gradient banding
-
 Row Level Security
 app.cogniba.com
-Landing page
 
 GitHub release
