+++ package.json
@@ -36,7 +36,7 @@
     "dotenv": "^16.4.5",
     "drizzle-orm": "^0.32.0",
     "lucide-react": "^0.411.0",
-    "next": "^14.2.15",
+    "next": "^14.2.17",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-confetti": "^6.1.0",
