+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
-// import "./blueGlobals.css";
-import "./orangeGlobals.css";
+// import "@/styles/globals/blueGlobals.css";
+import "@/styles/globals/orangeGlobals.css";
 
 import RootLayoutWrapper from "@/components/RootLayoutWrapper";
 
@@ -19,8 +19,8 @@ const fontSerif = Playfair_Display({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Cogniba",
+  description: "The only proven way to increase your intelligence.",
 };
 
 interface RootLayoutProps {
