+++ src/styles/globals/defaultBlueGlobals.css
