+++ src/styles/globals/orangeGlobals.css
@@ -26,7 +26,7 @@
     --popover: 255 255 255; /* #ffffff */
     --popover-foreground: 2 8 23; /* #020817 */
 
-    --primary: 37 99 235; /* #2563eb */
+    --primary: 234 88 12; /* #ea580c */
     --primary-foreground: 248 250 252; /* #f8fafc */
 
     --secondary: 241 245 249; /* #f1f5f9 */
@@ -44,15 +44,35 @@
     --successs: 34 197 94; /* #22c55e */
     --successs-foreground: 248 250 252; /* #f8fafc */
 
+    --link: 37 99 235; /* #2563eb */
+
     --border: 226 232 240; /* #e2e8f0 */
     --input: 226 232 240; /* #e2e8f0 */
-    --ring: 37 99 235; /* #2563eb */
+    --ring: 234 88 12; /* #ea580c */
 
     --chart-1: 231 110 80; /* #e76e50 */
     --chart-2: 42 157 144; /* #2a9d90 */
     --chart-3: 39 71 84; /* #274754 */
     --chart-4: 232 196 104; /* #e8c468 */
     --chart-5: 244 164 98; /* #f4a462 */
+
+    --red: 220 38 38; /* #dc2626 */
+    --orange: 234 88 12; /* #ea580c */
+    --amber: 217 119 6; /* #d97706 */
+    --yellow: 202 138 4; /* #ca8a04 */
+    --lime: 101 163 13; /* #65a30d */
+    --green: 22 163 74; /* #16a34a */
+    --emerald: 5 150 105; /* #059669 */
+    --teal: 13 148 136; /* #0d9488 */
+    --cyan: 8 145 178; /* #0891b2 */
+    --sky: 2 132 199; /* #0284c7 */
+    --blue: 37 99 235; /* #2563eb */
+    --indigo: 79 70 229; /* #4f46e5 */
+    --violet: 124 58 237; /* #7c3aed */
+    --purple: 147 51 234; /* #9333ea */
+    --fuchsia: 192 38 211; /* #c026d3 */
+    --pink: 219 39 119; /* #db2777 */
+    --rose: 225 29 72; /* #e11d48 */
   }
 
   .dark {
@@ -65,7 +85,7 @@
     --popover: 2 8 23; /* #020817 */
     --popover-foreground: 248 250 252; /* #f8fafc */
 
-    --primary: 59 130 246; /* #3b82f6 */
+    --primary: 249 115 22; /* #f97316 */
     --primary-foreground: 15 23 42; /* #0f172a */
 
     --secondary: 30 41 59; /* #1e293b */
@@ -83,15 +103,35 @@
     --success: 20 83 45; /* #14532d */
     --success-foreground: 248 250 252; /* #f8fafc */
 
+    --link: 59 130 246; /* #3b82f6 */
+
     --border: 30 41 59; /* #1e293b */
     --input: 30 41 59; /* #1e293b */
-    --ring: 29 78 216; /* #1d4ed8 */
+    --ring: 194 65 12; /* #c2410c */
 
     --chart-1: 38 98 217; /* #2662d9 */
     --chart-2: 46 184 138; /* #2eb88a */
     --chart-3: 232 140 48; /* #e88c30 */
     --chart-4: 175 87 219; /* #af57db */
     --chart-5: 226 54 112; /* #e23670 */
+
+    --red: 239 68 68; /* #ef4444 */
+    --orange: 249 115 22; /* #f97316 */
+    --amber: 245 158 11; /* #f59e0b */
+    --yellow: 234 179 8; /* #eab308 */
+    --lime: 132 204 22; /* #84cc16 */
+    --green: 34 197 94; /* #22c55e */
+    --emerald: 16 185 129; /* #10b981 */
+    --teal: 20 184 166; /* #14b8a6 */
+    --cyan: 6 182 212; /* #06b6d4 */
+    --sky: 14 165 233; /* #0ea5e9 */
+    --blue: 59 130 246; /* #3b82f6 */
+    --indigo: 99 102 241; /* #6366f1 */
+    --violet: 139 92 246; /* #8b5cf6 */
+    --purple: 168 85 247; /* #a855f7 */
+    --fuchsia: 217 70 239; /* #d946ef */
+    --pink: 236 72 153; /* #ec4899 */
+    --rose: 244 63 94; /* #f43f5e */
   }
 }
